python-consumers/requirements.txt
kafka-python==2.0.2
six